% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{arraytonc}
\alias{arraytonc}
\title{arraytonc: Create an nc file from an array}
\usage{
arraytonc(a, fileout, varname, units, r, tme, baseyear = 1900)
}
\arguments{
\item{a}{the array containing the data to be used to create the new netCDF file.}

\item{fileout}{the name and location of the nteCDF file to be created.}

\item{varname}{the name of the variable to be created (character string).}

\item{units}{The variable's units (character string). Or, pass a zero length string (") to have no units attribute.}

\item{r}{a raster file with extent parameters that the new netCDF file should follow.}

\item{tme}{an object of class \code{POSIXlt} representing calendar dates and times.}

\item{baseyear}{the calendar year that measurements began.}
}
\value{
an object of class ncdf4 that has the fields described above and is saved to the location specified by 'fileout'.
}
\description{
\code{arraytonc} creates a new netCDF file from an array object.
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.

temps <- array(10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460), dim=c(180,360,1460))
r <- raster()
tme <- tmecreate(2010, 6)
arraytonc (temps, "new.nc", varname = "6-hourly air temperature", units = "degrees Celcius", r, tme, baseyear = 1900)
}
