% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio1.R
\name{bio1}
\alias{bio1}
\title{bio1: Calculate mean annual temperature.}
\usage{
bio1(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value. Ignored if method unspecified.}

\item{method}{an optional character string describing the method used to
calculate mean annual temperature. Options are "dailymaxmin" or unspecified
(see details).}
}
\value{
a single numeric value of mean annual temperature.
}
\description{
\code{bio1} is used to calculate mean annual temperature.
}
\details{
If \code{method} is "dailymaxmin", daily mean temperatures are
calculated as the mean of daily maximum and minimum temperatures and annual
mean calculated from daily means. Otherwise the mean of \code{temps} is returned.
If using \code{dailymaxmin} method and data span more than one year, calculations
will be performed on all data and a single value returned.
}
\examples{
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio1(hourly_temps, tme)
bio1(hourly_temps, tme, method = "dailymaxmin")
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
