% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio10.R
\name{bio10}
\alias{bio10}
\title{bio10: Calculate mean temperature of the warmest quarter.}
\usage{
bio10(temps, tme)
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}
}
\value{
a single numeric value of mean temperature in the warmest quarter of
the year.
}
\description{
\code{bio10} is used to calculate the mean temperature of the warmest
quarter of the year.
}
\details{
Mean temperature of the warmest 3-month (91-day) period is
calculated from annual temperature values. If data span more than one year,
calculations are performed on all data and a single value returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio10(temps, tme)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
