% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio12.R
\name{bio12}
\alias{bio12}
\title{bio12: Calculate total annual precipitation.}
\usage{
bio12(prec, tme)
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}
}
\value{
a single numeric value of total annual precipitation.
}
\description{
\code{bio12} is used to calculate total precipitation in the year.
}
\details{
All precipitation values for each year are summed. If data span more
than one year, calculations are performed on all data and a single value
returned.
}
\examples{
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Precipitation")
bio12(prec, tme)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
