% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio13.R
\name{bio13}
\alias{bio13}
\title{bio13: Calculate precipitation of the wettest period.}
\usage{
bio13(prec, tme, period = 7)
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}

\item{period}{a single value defining the number of days for which total
precipitation is calculated for.}
}
\value{
a single numeric value of total precipitation in the wettest period
of the year.
}
\description{
\code{bio13} is used to calculate the total precipitation in the
wettest period of the year.
}
\details{
The total precipitation in a given period starting at each data
point is calculated, and the wettest returned. By default the period is set
to 7 days. If data span more than one year, data across the whole period are
considered and a single value returned.
}
\examples{
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Precipitation")
bio13(prec, tme, period = 7)
bio13(prec, tme, period = 30)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
