% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio15.R
\name{bio15}
\alias{bio15}
\title{bio15: Calculate precipitation seasonality.}
\usage{
bio15(prec, tme)
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}
}
\value{
a single numeric value representing annual precipitation seasonality.
}
\description{
\code{bio15} is used to calculate the differences in mean
precipitation between the wettest and driest consecutive three month (90 day)
period.
}
\details{
Each provided precipitation data point is treated as the start of a
90 day period for which mean temperature is calculated. The differences
between consecutive 90 day means are then calculated, and the greatest
absolute difference returned.
}
\examples{
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Precipitation")
bio15(prec, tme)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
