% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio17.R
\name{bio17}
\alias{bio17}
\title{bio17: Calculate precipitation of the driest quarter.}
\usage{
bio17(prec, tme)
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}
}
\value{
a single numeric value of total precipitation of the driest quarter.
}
\description{
\code{bio17} is used to calculate the total precipitation in the
driest quarter of the year
}
\details{
Precipitation in quarter is calculated and total
precipitation in the driest quarter returned. If data span more than one
year, calculations are performed on all data and single value returned.
}
\examples{
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Precipitation")
bio17(prec, tme)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
