% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio18.R
\name{bio18}
\alias{bio18}
\title{bio18: Calculate precipitation of the warmest quarter.}
\usage{
bio18(temps, prec, tme1, tme2)
}
\arguments{
\item{temps}{a vector of temperature values, normally for one year (see
details).}

\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}
}
\value{
a single numeric value of total precipitation in the warmest quarter
of the year.
}
\description{
\code{bio18} is used to calculate the total precipitation in the
warmest quarter of the year.
}
\details{
If \code{temps} and \code{prec} vary in their temporal resolution, then they
are harmonised by aggregation (mean for temperature or sum for precipitation).
Then, each data point is treated as the start of a 91 day period for which
summed temperature is calculated. Once the warmest 91 day period is
identified, the total precipitation for the same period is calculated and a
single value returned. If data span more than one year, calculations are
performed on all data and a single value returned.
}
\examples{
# hourly
tme1 <- tmecreate(2019, 1)
# 6-hourly
tme2 <- tmecreate(2019, 6)
# daily
tme3 <- tmecreate(2019, 24)
plot(hourly_precip~as.POSIXct(tme1), type = "l", xlab = "Month",
ylab = "Precipitation")
bio18(hourly_temps, hourly_precip, tme1, tme2)
bio18(six_hourly_temps, daily_precip, tme2, tme3)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
