% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio2.R
\name{bio2}
\alias{bio2}
\title{bio2: Calculate mean annual diurnal temperature range.}
\usage{
bio2(temps, tme, min_percentile = 0.01, max_percentile = 0.99)
}
\arguments{
\item{temps}{a vector of sub-daily temperatures, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}

\item{min_percentile}{a value between 0-1 used to define the minimum
percentile.}

\item{max_percentile}{a value between 0-1 used to define the maximum
percentile.}
}
\value{
a single numeric value of mean diurnal temperature range.
}
\description{
\code{bio2} is used to calculate the mean annual diurnal range in
temperature (range mean of the maximum-minimum).
}
\details{
The daily maximum (percentile dependant) and minimum (percentile
dependant) temperatures are calculated. The mean of the difference between
both is returned. By default, percentiles are set at 1\% and 99\% for the
minimum and maximum values respectively. If data spans more than one
year then a single value considering all data is returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio2(temps, tme)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
