% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio3.R
\name{bio3}
\alias{bio3}
\title{bio3: Calculate isothermality.}
\usage{
bio3(temps, tme, min_percentile = 0.01, max_percentile = 0.99)
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}

\item{min_percentile}{a value between 0-1 used to define the minimum
percentile.}

\item{max_percentile}{a value between 0-1 used to define the maximum
percentile.}
}
\value{
a single numeric value representing isothermality for each year.
}
\description{
\code{bio3} is used to calculate isothermality (day-to-night
temperature oscillations relative to annual oscillations).
}
\details{
bio3 is calculated using the mean of daily maximum and minimum
temperature. If data spans more than one year, data are aggregated by unique
month irrespective of year and one value returned.
}
\examples{
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio3(hourly_temps, tme)
bio3(hourly_temps, tme, min_percentile = 0.05, max_percentile = 0.95)
}
\seealso{
\code{\link[=bio2]{bio2()}} and \code{\link[=bio7]{bio7()}} for calculating diurnal and annual
temperature ranges. \code{\link[=tmecreate]{tmecreate()}} for creating a 'POSIXlt' object.
}
