% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio4.R
\name{bio4}
\alias{bio4}
\title{bio4: Calculate temperature seasonality.}
\usage{
bio4(temps, tme)
}
\arguments{
\item{temps}{a vector of temperature values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}
}
\value{
a single numeric value representing annual temperature seasonality.
}
\description{
\code{bio4} is used to calculate the differences in mean temperature
between the warmest and coldest consecutive three month (91 day) period.
}
\details{
Each provided temperature data point is treated as the start of a 91
day period for which mean temperature is calculated. The differences between
consecutive 91 day means are then calculated, and the greatest absolute
difference returned.
}
\examples{
# hourly
tme1 <- tmecreate(2010, 1)
# 6-hourly
tme2 <- tmecreate(2010, 6)
# daily
tme3 <- tmecreate(2010, 24)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio4(hourly_temps, tme)
bio4(six_hourly_temps, tme2)
bio4(daily_temps, tme3)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
