% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio5.R
\name{bio5}
\alias{bio5}
\title{bio5: Calculate maximum temperature of the year.}
\usage{
bio5(temps, tme, percentile = 0.99)
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}

\item{percentile}{a value between 0-1 used to define the maximum percentile.}
}
\value{
a single numeric value of maximum temperature for a year.
}
\description{
\code{bio5} is used to calculate the maximum temperature of the year.
}
\details{
The maximum temperature (percentile dependant) across all values is
returned. By default, percentile is set to 99\%. If data spans more than one
year then a single value considering all data is returned.
}
\examples{
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio5(hourly_temps, tme)
bio5(hourly_temps, tme, percentile = 0.95)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
