% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio6.R
\name{bio6}
\alias{bio6}
\title{bio6: Calculate minimum temperature of the year.}
\usage{
bio6(temps, tme, percentile = 0.01)
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}

\item{percentile}{a value between 0-1 used to define the minimum percentile.}
}
\value{
a single numeric value of minimum temperature for a year.
}
\description{
\code{bio6} is used to calculate the minimum temperature of the year.
}
\details{
The minimum temperature (percentile dependant) across all values is
returned. By default, percentile is set to 1\%. If data spans more than one
year then a single value considering all data is returned.
}
\examples{
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio6(hourly_temps, tme)
bio6(hourly_temps, tme, percentile = 0.05)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
