% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consec.R
\name{consec}
\alias{consec}
\title{consec: Calculate consecutive measurements below or above a specified
threshold.}
\usage{
consec(x, threshold, below = TRUE)
}
\arguments{
\item{threshold}{a value with which to threshold the data in \code{var}.}

\item{below}{a logical indicating whether values above or below the threshold
are to be considered. Default = TRUE.}

\item{var}{a vector of values for which a threshold is to be applied.}
}
\value{
a single numeric value of the longest run of measurements that adhere
to the threshold criteria.
}
\description{
\code{consec} is used to find consecutive values above or below
a user-defined threshold.
}
\details{
???
}
\examples{
# calculate waterlogged hours
set.seed(100)
rain <- rgamma(365,4) * rbinom(365,1,0.3)
evap <- rep(1.15,365)
cn <- 90
s_min <- 0.074
s_max <- 0.422
s_depth <- 0.5
sm <- soil_moisture(rain, evap, cn)
wlh <- consec(sm, threshold = 0.422, below = FALSE)
}
