% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{cumsumseq}
\alias{cumsumseq}
\title{cumsumseq
#'@description \code{cumsumseq} is used to create a vector of values indicating growing season conditions for the specified time period.}
\usage{
cumsumseq(gs, ehour, div)
}
\arguments{
\item{gs}{a vector of binary values indicating growing season (1) or non-growing season (0).}

\item{ehour}{a vector of evapotranspiration values.}

\item{div}{numeric object specifying the value by which to divide total days of each month.}
}
\value{
a vector of numeric values alternating between 0 and 1.
}
\description{
cumsumseq
#'@description \code{cumsumseq} is used to create a vector of values indicating growing season conditions for the specified time period.
}
\examples{
gs <- c(0,0,1,1,1,1,0,0,1,1,0)
ehour <- c(80, 81, 82, 83, 84, 85, 84, 83, 82, 81, 80)
div = 2
cs <- cumsumseq(gs, ehour, div = 2)

}
