% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_pev}
\alias{daily_pev}
\title{Daily potential evaporation for 2019.}
\format{A numeric vector of potential evaporation (mm/day).}
\source{
\url{https://www.ecmwf.int/en/era5-land}
}
\usage{
daily_pev
}
\description{
A dataset containing daily potential evaporation for 2019 from the
ERA-5 Land dataset. Data for a location near Kinlochewe, Scotland (-5.2,
57.6).
}
\keyword{datasets}
