% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_temps}
\alias{daily_temps}
\title{Daily temperatures for 2019.}
\format{A numeric vector of temperatures (degrees ceclius).}
\source{
\url{https://www.ecmwf.int/en/era5-land}
}
\usage{
daily_temps
}
\description{
A dataset containing daily 2m air temperature for 2019 from
the ERA-5 Land dataset. Data for a location near Kinlochewe, Scotland
(-5.2, 57.6).
}
\keyword{datasets}
