% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff.R
\name{.SCN_adjust}
\alias{.SCN_adjust}
\title{Adjusts run-off curve number by antecedent soil moisture}
\usage{
.SCN_adjust(scn, sm, Smin, Smax)
}
\description{
Adjusts run-off curve number by antecedent soil moisture
}
