% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{events}
\alias{events}
\title{events: determines the number of events}
\usage{
events(x)
}
\arguments{
\item{x}{a vector with 0, indicating non-event (e.g. no rainfall), and any other value indicating an event.}
}
\value{
an integer representing the total number of events.
}
\description{
\code{events} is used to calculate the total number of events (e.g. number of rainfall events).
}
\examples{
prec <- sample(rep(0:10,10))
events(prec)

}
