% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_autumn_frost.R
\name{first_autumn_frost}
\alias{first_autumn_frost}
\title{first_autumn_frost: Identify date of first Autumn frost.}
\usage{
first_autumn_frost(temps, tme, northern_hemisphere = TRUE)
}
\arguments{
\item{temps}{a vector of temperatures, spanning a year.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}

\item{northern_hemisphere}{a logical indicating whether Spring dates are to
be calculated for the northern (TRUE) or southern (FALSE) hemisphere.}
}
\value{
a POSIXlt object representing the date of the first Autumn frost, or
NA if no frost occurred.
}
\description{
\code{first_autumn_frost} is used to find the date of the first
Autumn frost.
}
\details{
Autumn dates are determined based on the hemisphere chosen. For the
northern hemisphere Autumn is assumed to begin 1st March and end on 31st May.
For the southern hemisphere, Spring is assumed to begin on 1st September and
finish on 30th November. If no frosts occurred within the Autumn period, NA
is returned.
}
\examples{
temps <- hourly_temps - 5
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Temperature")
first_autumn_frost(temps, tme)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
