% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frost_degree_hours.R
\name{frost_degree_hours}
\alias{frost_degree_hours}
\title{frost_degree_hours: Calculate frost degree hours.
#'}
\usage{
frost_degree_hours(temps, tme)
}
\arguments{
\item{temps}{a vector of hourly temperatures.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}
}
\value{
a single numeric value of frost degree hours.
}
\description{
\code{frost_hours} is used to calculate the number of frost hours for
a given period of time.
}
\details{
For each provided hourly temperature value, the number of degrees
celcius below 0 are calculated. These values are then summed to provide the
frost degree hours over a given period of time.
}
\examples{
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Temperature")
frost_degree_hours(hourly_temps, tme)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
