% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frost_free_period.R
\name{frost_free_period}
\alias{frost_free_period}
\title{frost_free_period: Calculate the length of the longest frost free period.}
\usage{
frost_free_period(temps, tme, period = "year", pstart = NA, pend = NA)
}
\arguments{
\item{temps}{a vector of temperatures, spanning a year.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}

\item{period}{character vector indicating either year or custom.}

\item{pstart}{a \code{POSIXlt} object representing the date and time of the start
of a user-defined period. Only used if period = "custom".}

\item{pend}{a \code{POSIXlt} object representing the date and time of the start of
a user-defined period. Only used in period = "custom".}
}
\value{
a single numeric value indicating the number of time steps (as
determined by \code{tme}) that are frost free.
}
\description{
\code{frost_free_period} is used to find the length of a frost free
period, for a given season or yearly.
}
\details{
values are returned at the time step of input data.
}
\examples{
temps <- hourly_temps - 5
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Temperature")
frost_free_period(temps, tme)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
