% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frost_hours.R
\name{frost_hours}
\alias{frost_hours}
\title{frost_hours: Calculate frost hours.}
\usage{
frost_hours(temps, tme)
}
\arguments{
\item{temps}{a vector of hourly temperatures.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}
}
\value{
a single numeric value of frost hours.
}
\description{
\code{frost_hours} is used to calculate the number of frost hours for
a given period of time.
}
\details{
Frost hours are defined as temperature values that fall below 0
degrees celcius.
}
\examples{
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Temperature")
frost_hours(hourly_temps, tme)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
