% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseason.R, R/physclimvars.R
\name{gseason}
\alias{gseason}
\title{gseason: Calculate the period when plant growth is possible.}
\usage{
gseason(
  year,
  precipnc,
  evapnc,
  tempnc,
  period = "monthly",
  surplus = TRUE,
  lower = 5,
  upper = 35,
  nday = 1/24,
  daynight = TRUE,
  merid = 0,
  dst = 0
)

gseason(
  year,
  precipnc,
  evapnc,
  tempnc,
  period = "monthly",
  surplus = TRUE,
  lower = 5,
  upper = 35,
  nday = 1/24,
  daynight = TRUE,
  merid = 0,
  dst = 0
)
}
\arguments{
\item{year}{calendar year.}

\item{precipnc}{nc file containing precipitation values.}

\item{evapnc}{nc file containing evpotranspiration values.}

\item{tempnc}{nc file containing temperature values.}

\item{period}{an optional character string defining time period of precipitation and evapotranspiration measurements. Options include "monthly" and "bimonthly". If left unspecified, assumed 'submonthly'.}

\item{surplus}{either TRUE or FALSE.}

\item{lower}{defines lower temperature limit where plant growth ceases (degrees Celcius) and growing season is switched off.}

\item{upper}{defines upper temeprature limit where plant growth ceases (degrees Celcius) and growing season is switched off.}

\item{nday}{specifies the number of consecutive days of temperatures above 'lower' or below 'upper' before growing season start/end is accepted.}

\item{daynight}{if TRUE, growing season is continuous over 24 hours, if FALSE, only calculates growing season during the day.}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is 0.}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0). Default is 0.}

\item{temps}{a vector of temperature values.}

\item{prec}{a vector of precipitation values.}

\item{evap}{a vector of evapotranspiration values.}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}

\item{tme3}{a \code{POSIXlt} object representing the date and time of each \code{evap}
value.}

\item{lon}{a numeric value indicating the longitudinal position of data (in
decimal degrees).}

\item{lat}{a numeric value indicating the latitudinal position of data (in
decimal degrees).}
}
\value{
an array of binary values indicating growing conditions (1) = yes,
(0) = no.

a three dimensional array of binary values indicating growing conditions (1) = yes, (0) = no.
}
\description{
\code{gseason} calculates total annual growing hours where
temperatures are within defined upper and lower limits and precipitation
exceeds half potential evapotranspiration.

\code{gseason} calculates total annual growing hours where temperatures are within defined upper and lower limits and precipitation exceeds half potential evapotranspiration.
}
\details{
The growing season period is defined as the period where
temperatures are >5 degree Celcius and <35 degree Celcius and precipitation0.5 PET.

The growing season period is defined as the period where temperatures are >5 degree Celcius and <35 degree Celcius and precipitation > 0.5 PET.
}
\seealso{
\code{\link[=gseason_prec]{gseason_prec()}} calculates period where precipitation exceeds half
evapotranspiration.

\code{\link[=gseason_temp]{gseason_temp()}} calculates period where temperature are above lower
limit for plant growth and below upper limit for plant growth.

\code{\link[=gseason_day]{gseason_day()}} calculates day/night time.

\code{\link[=tmecreate]{tmecreate()}} can be used to create a POSIXlt object.

temps <- hourly_temps
prec <- hourly_prec
evap <- prec + rnorm(length(prec)) - rnorm(length(prec))

tme1 <- tmecreate(2010,1)
tme2 <- tmecreate(2010,1)
tme3 <- tmecreate(2010,1)
lon <- -5
lat <- 52
out <- gseason(temps, prec, evap, tme1, tme2, tme3, lon, lat)

\code{\link[=gseason_prec]{gseason_prec()}} calculates period where precipitation exceeds half evapotranspiration.

\code{\link[=gseason_temp]{gseason_temp()}} calculates period where temperature are above lower limit for plant growth and below upper limit for plant growth.

\code{\link[=gseason_day]{gseason_day()}} calculates day/night time.

\code{\link[=tmecreate]{tmecreate()}} can be used to create a POSIXlt object.

\code{\link[=nctoarray]{nctoarray()}} can be used to create an nc file from array output.
}
