% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gseason_day}
\alias{gseason_day}
\title{gseason_day: Determine daytime hours}
\usage{
gseason_day(tme, lat, long, merid = 0, dst = 0)
}
\arguments{
\item{tme}{a \code{POSIXlt} object.}

\item{lat}{latitude of the location for which gseason_day is required.}

\item{long}{longitude of the location for which gseason_day is required.}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is 0.}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0). Default is 0.}
}
\value{
a vector of binary values indicating day (1) or night (0).
}
\description{
\code{gseason_day} is used to calculate daytime (1) and non-daytime (0) hours.
}
\examples{
Daytime hours in Porthleven, Cornwall for 2010.
tme <- tmecreate(2010, 1)
gseason_day(tme, 6, 21)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.

Requires wrapper function \code{sisimple} to be loaded.
}
