% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseason_length.R
\name{gseason_length}
\alias{gseason_length}
\title{gseason_length: Calculate length of growing season.}
\usage{
gseason_length(gs, tme, unit = "day")
}
\arguments{
\item{gs}{a vector of binary values indicating growing conditions
(1) = yes, (0) = no.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{gs} value.}

\item{unit}{a character vector, either "day" or "hour" defining the desired
temporal resolution of the value returned.}
}
\value{
a single numeric value indicating the length of the growing season
to the nearest hour or day depending on \code{unit}.
}
\description{
\code{gseason_length} calculates the length of the growing season
period.
}
\details{
If \code{unit} = "day" and data are a finer temporal resolution, then
the returned value is rounded to the nearest day.
}
\examples{
tme <- tmecreate(2010,1)
lon <- -5
lat <- 52
gs <- gseason(hourly_temps, hourly_precip, hourly_pev, tme, tme, tme, lon,
lat)

gsl <- gseason_length(gs, tme, unit = "day")
gsl <- gseason_length(gs, tme, unit = "hour")

}
\seealso{
the \code{\link[=gseason]{gseason()}} function can be used to create an array of growing
conditions (1) = yes, (0) = no accounting for temperature, precipitation and
daylight hours.
}
