% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseason_temp.R
\name{gseason_temp}
\alias{gseason_temp}
\title{gseason_temp: Calculate the growing season period as determined by
temperature values.}
\usage{
gseason_temp(temps, tme, lower = 5, upper = 35, nday = 5)
}
\arguments{
\item{temps}{a vector of temperature values.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}

\item{lower}{integer specifying lower temperature limit of growing season.}

\item{upper}{integer specifying upper temperature limit of growing season.}

\item{nday}{specifies the number of consecutive days of temperatures above
'lower' or below 'upper' before growing season start/end is accepted.}
}
\value{
A time-series object of binary values, where 1 indicates growing
season and 0 indicates not growing season.
}
\description{
\code{gseason_temp} calculates growing season period where
temperatures are above 'lower' and below 'upper' specified limits.
}
\details{
To satisfy the requirements for \code{tme}, a POSIXlt object can be
created using the \code{tmecreate} wrapper function.
}
\examples{
tme <- tmecreate(2010, 1)
gseast <- gseason_temp(temps = hourly_temps, tme)
plot(gseast)

}
\seealso{
\code{\link[=tmecreate]{tmecreate()}} can be used to create a POSIXlt object.
}
