% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gseasonlength}
\alias{gseasonlength}
\title{gseasonlength: Growing season length}
\usage{
gseasonlength(startyear, endyear, gsnc, msk, fo)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{gsnc}{full path name of nc file containing growing season values for each year (1 = growing season, 0 = non-growing season).}

\item{msk}{optional land/sea mask raster object with sea areas assigned 'NA' and land areas assigned value '1'.}

\item{fo}{character string indicating the name and file path for the .tif file to be saved.}
}
\value{
A raster object of values of growing season length (number of days) exported as a .tif file
}
\description{
\code{gseasonlength} calculates the average length of the growing season period (in days) across specified years.
}
\details{
Function has been designed to run with raster data of dimensions 73 x 144.
}
\seealso{
the \code{gseason} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.

\code{mtoraster} to convert a matrix to a raster object.

\code{nctoarray} to create an array of growing season values from an nc file.

Requires function \code{gsl} to be loaded.
}
