% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gseastemp}
\alias{gseastemp}
\title{gseastemp: Mean temperature during the growing season}
\usage{
gseastemp(startyear, endyear, temp, gseason)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{temp}{array of temperature values for one year.}

\item{gseason}{array of binary values indicating growing (1) or non-growing (0) season for one year.}
}
\value{
A matrix of mean growing season temperature values over the specified years.
}
\description{
\code{gseastemp} calculates the mean air temperature during the growing season over specified years.
}
\details{
details to be added
}
\seealso{
the \code{\link[=gseason]{gseason()}} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.

Requires function \code{\link[=gst]{gst()}} to be loaded.

\code{\link[=mtoraster]{mtoraster()}} to convert output matrix to a raster.

\code{\link[=nctarray]{nctarray()}} to create an array from nc file.
}
