% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gsl}
\alias{gsl}
\title{gsl: Growing season length}
\usage{
gsl(gseason, year)
}
\arguments{
\item{gseason}{a three dimensional array of binary values (1 = growing season, 0 = not growing season).}

\item{year}{calander year.}
}
\value{
a matrix of values of growing season length in decimal days.
}
\description{
\code{gsl} calculates the length of the growing season in decimal days.
}
\examples{
tme <- as.POSIXlt(c(0:1459) * 3600 * 6, origin = "2010-01-01 00:00", tz = "GMT")
gs <- gseason_day(tme, 6, 21)
gsa <- array(gs, dim=c(1, 1, 1460))
gsl <- gsl(gsa, 2010)

}
\seealso{
the \code{gseason} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.
}
