% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsp.R
\name{gsp}
\alias{gsp}
\title{gsp: Calculate total precipitation during growing season.}
\usage{
gsp(prec, gs)
}
\arguments{
\item{prec}{a vector of precipitation values.}

\item{gs}{a vector of binary values indicating growing conditions
(1) = yes, (0) = no.}
}
\value{
a single numeric value of the total precipitation during growing
conditions.
}
\description{
\code{gsp} calculates total precipitation during the growing season.
}
\examples{
temps <- hourly_temps
prec <- hourly_prec
evap <- prec + rnorm(length(prec)) - rnorm(length(prec))

tme1 <- tmecreate(2010,1)
tme2 <- tmecreate(2010,1)
tme3 <- tmecreate(2010,1)
lon <- -5
lat <- 52
gs <- gseason(temps, prec, evap, tme1, tme2, tme3, lon, lat)
totalgsp <- gsp(prec, gs)

}
\seealso{
the \code{\link[=gseason]{gseason()}} function can be used to create an array of growing
conditions (1) = yes, (0) = no accounting for temperature, precipitation and
optionally daylight hours.
}
