% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssm.R
\name{gssm}
\alias{gssm}
\title{gssm: Calculate mean soil moisture during the growing season.}
\usage{
gssm(sm, gs)
}
\arguments{
\item{sm}{a vector of soil moisture values (e.g. calculated by
\code{soil_moisture}).}

\item{gs}{a vector of binary values indicating growing conditions
(1) = yes, (0) = no.}
}
\value{
a single numeric value of the mean temperature during growing
conditions.
}
\description{
\code{gssm} calculates mean temperature during the growing season.
}
\details{
\code{sm} must be the same length as \code{gseason}.
}
\examples{
tme <- tmecreate(2010, 1)
evap <- rep(1.15,length(tme))
sm <- soil_moisture(hourly_precip, evap, 0.9)
gssm(sm, gs)
}
\seealso{
the \code{\link[=gseason]{gseason()}} function can be used to create an array of growing
conditions (1) = yes, (0) = no accounting for temperature, precipitation and
optionally daylight hours.
}
