% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gst.R
\name{gst}
\alias{gst}
\title{gst: Calculate mean temperature during the growing season.}
\usage{
gst(temps, gs)
}
\arguments{
\item{temps}{a vector of temperature values.}

\item{gs}{a vector of binary values indicating growing conditions
(1) = yes, (0) = no.}
}
\value{
a single numeric value of the mean temperature during growing
conditions.
}
\description{
\code{gst} calculates mean temperature during the growing season.
}
\details{
\code{temps} must be the same length as \code{gseason}.
}
\examples{
temps <- hourly_temps
prec <- hourly_prec
evap <- prec + rnorm(length(prec)) - rnorm(length(prec))

tme1 <- tmecreate(2010,1)
tme2 <- tmecreate(2010,1)
tme3 <- tmecreate(2010,1)
lon <- -5
lat <- 52
gs <- gseason(temps, prec, evap, tme1, tme2, tme3, lon, lat)
meangst <- gst(temps, gs)
}
\seealso{
the \code{\link[=gseason]{gseason()}} function can be used to create an array of growing
conditions (1) = yes, (0) = no accounting for temperature, precipitation and
optionally daylight hours.
}
