% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hourly_pev}
\alias{hourly_pev}
\title{Hourly potential evaporation for 2019.}
\format{A numeric vector of potential evaporation (mm/hour).}
\source{
\url{https://www.ecmwf.int/en/era5-land}
}
\usage{
hourly_pev
}
\description{
A dataset containing hourly potential evaporation for 2019 from the
ERA-5 Land dataset. Data for a location near Kinlochewe, Scotland (-5.2,
57.6).
}
\keyword{datasets}
