% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_spring_frost.R
\name{last_spring_frost}
\alias{last_spring_frost}
\title{last_spring_frost: Identify date of last Spring frost.}
\usage{
last_spring_frost(temps, tme, northern_hemisphere = TRUE)
}
\arguments{
\item{temps}{a vector of temperatures, spanning a year.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}

\item{northern_hemisphere}{a logical indicating whether Spring dates are to
be calculated for the northern (TRUE) or southern (FALSE) hemisphere.}
}
\value{
a POSIXlt object representing the date of the last Spring frost, or
NA if no frost occurred.
}
\description{
\code{last_spring_frost} is used to find the date of the last Spring
frost.
}
\details{
Spring dates are determined based on the hemisphere chosen. For the
northern hemisphere Spring is assumed to begin 1st March and end on 31st May.
For the southern hemisphere, Spring is assumed to begin on 1st September and
finish on 30th November. If no frosts occurred within the Spring period, NA
is returned.
}
\examples{
temps <- hourly_temps - 5
tme <- tmecreate(2010, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Temperature")
last_spring_frost(temps, tme)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
