% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{mst}
\alias{mst}
\title{mst: Calculate mean summer temperature.}
\usage{
mst(temps, tme, northern_hemisphere = TRUE)
}
\arguments{
\item{temps}{a vector of temperature values.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{prec} value.}

\item{northern_hemisphere}{a logical indicating whether Spring dates are to
be calculated for the northern (TRUE) or southern (FALSE) hemisphere.}
}
\value{
a single numeric value of the mean temperature during summer.
}
\description{
\code{mst} calculates the mean temperature during summer, accounting
for differences in the timing of the summer period for the northern and
southern hemispheres.
}
\details{

}
\examples{
tme <- tmecreate(2010, 1)
mst(hourly_temps, tme)
mst(hourly_temps, tme, northern_hemisphere = FALSE)
}
