% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{mtoraster}
\alias{mtoraster}
\title{mtoraster: Convert a matrix to raster object}
\usage{
mtoraster(m, centre = TRUE, mask = NA)
}
\arguments{
\item{m}{a matrix of values.}

\item{centre}{if true, centres values to prime meridian (longitude 0°).}

\item{mask}{option to mask land/sea from plotted values.}
}
\value{
a raster object.
}
\description{
\code{mtoraster} converts a matrix of global values to raster format.
}
\details{
the function is written for the conversion of matrices of dimensions 73 x 144.
}
\examples{
m <- matrix(rnorm(0:72), 73, 144)
r <- mtoraster(m)

}
