% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{nctoarray}
\alias{nctoarray}
\title{nctoarray: Creates an array containing data within an nc file}
\usage{
nctoarray(filein, varid = NA)
}
\arguments{
\item{filein}{A character string with the full path name of the nc file. Tilde-expansion is performed.}

\item{varid}{The variable to be read from the ncfile. Can be a string with the
name of the variable or an object of class ncvar4. If left unspecified, the
function will determine if there is only one variable in the file and, if
so, read from that. If left unspecified and there are multiple variables in
the file, an error is generated.}
}
\value{
a three-dimensional array of the specified variable, permuted to dimension order 2, 1, 3.
}
\description{
\code{nctoarray} is used to create a three-dimensional array
containing data from a .nc file.
}
\examples{
to follow once there is an nc file in data
# ========= Create nc file of random data ============================= #
longs <- ncdim_def(name="lon", units = "degrees_east", vals = seq(1,10,1),longname="longitude")
lats <- ncdim_def(name="lat",  units = "degrees_north",vals=seq(1,10,1),longname="latitude")
times <- ncdim_def(name="time",units="hours since 1983-01-01 00:00:00", vals=c(101:110))
mydata <-ncvar_def(name="Random data", units="mm", dim = list(longs, lats, times))
ncnew <- nc_create(filename="Random.nc", mydata)
ncvar_put(ncnew, mydata, vals = array(rnorm(10000), dim = c(10,10,10)))
nc_close(ncnew)

# ========= Read in as array ============================= #
arr <- nctoarray("Random.nc", varid = "Random data")

}
