% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{resamplearray}
\alias{resamplearray}
\title{resamplearray: resamples an array to the dimensions of another array}
\usage{
resamplearray(a, rin, rout)
}
\arguments{
\item{a}{the array object to be resampled.}

\item{rin}{a raster object used to resample extent of 'a' to extent of 'rout'.}

\item{rout}{a raster object with extent parameters that 'a' should be resampled to.}
}
\value{
an array with the same latitude and longitude as 'rout'.
}
\description{
reformats data in an array and creates a new array objects to
dimensions specified by \code{rin}.
}
\examples{
a <- array(rnorm(100), dim=c(94,192,1464))
rin <- raster(a[,,1])
extent(rin) <- c(-0.9375, 359.0625, -89.49406, 89.49406)
routm <- matrix(rnorm(10512), 73, 144)
rout <- raster(routm)
extent(rout) <- c(-1.25, 358.75, -91.25, 91.25)
newarray <- resamplearray(a, rin, rout)
dim(newarray)

}
