% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{resamplenc}
\alias{resamplenc}
\title{resamplenc: Reformat an nc file to new dimensions.}
\usage{
resamplenc(filein1, filein2, varname, units, fileout)
}
\arguments{
\item{filein1}{master nc file.}

\item{filein2}{nc file to undergo reformatting.}

\item{varname}{the name of the variable to be created (character string).}

\item{units}{the variable's units (character string). Or, pass a zero length string (") to have no units attributed.}

\item{fileout}{full path name of the new nc file to be created.}

\item{var}{time dimension to be left unchanged.}
}
\value{
a .nc file with the same dimensions as filein1.
}
\description{
\code{resamplenc} is used to reformat data an nc file to the same dimensions as the data in \code{filein1}.
}
\details{
recording period for filein2 is left unchanged.
}
\examples{
to follow once there are two nc files with different latitude and longitude dimensions in the data.

}
\seealso{
\code{nctoarray}
}
