% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff.R
\name{runoff}
\alias{runoff}
\title{Calculates runoff.}
\usage{
runoff(rain, cn, sm, s_min, s_max)
}
\arguments{
\item{rain}{rainfall in mm.}

\item{cn}{run-off curve number (see details).}

\item{sm}{soil water fraction.}

\item{s_min}{residual water content (m^3 / m^3).}

\item{s_max}{volumetric water content at saturation (m^3 / m^3).}
}
\value{
direct runoff in mm.
}
\description{
\code{runoff} calculates direct runoff based on USDA Run-off curve
number.
}
\details{
Applies the United States Department of Agriculture method for
calculating direct runoff. Runoff curve numbers can be obtained from here:
https://www.wcc.nrcs.usda.gov/ftpref/wntsc/H&H/training/runoff-curve-numbers1.pdf
Adjustments for antecedent moisture condition are automatically applied, so
the supplied value of \code{cn} should be for average soil moisture (AMC II).
\code{s_min} and \code{s_max} values for soil types are given in \code{\link[=soilparams]{soilparams()}}.
}
\examples{
rain <- c(1:150)
ro1 <- runoff(rain, cn = 50, sm = 0.4, s_min = 0.074, s_max = 0.422)
ro2 <- runoff(rain, cn = 90, sm = 0.4, s_min = 0.074, s_max = 0.422)
plot(ro1 ~ rain, type = "l")
plot(ro2 ~ rain, type = "l")
}
