% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sisimple}
\alias{sisimple}
\title{sisimple: Calculates the solar index}
\usage{
sisimple(localtime, lat, long, julian, merid = 0, dst = 0)
}
\arguments{
\item{lat}{latitude of the location for which the solar index is required (decimal degrees, -ve south of the equator).}

\item{long}{longitude of the location for which the solar index is required (decimal degrees, -ve west of Greenwich meridian).}

\item{julian}{Julian day expressed as an integer as returned by julday().}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is 0.}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0).}

\item{locatime}{local time (decimal hour, 24 hour clock).}
}
\value{
a numeric value of the proportion of direct beam radiation on a horizontal surface at the specified latitude and longitude.
}
\description{
Calculates the proportion of direct beam radiation incident on an inclined surface at a specified time and location.
}
\examples{
jd <- julday (2010, 6, 21)
si <- sisimple(11, 50, -5, jd, merid=0, dst=0)

}
\seealso{
the microclima function \code{julday} can be used to derive \code{julian}.
}
