% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{six_hourly_temps}
\alias{six_hourly_temps}
\title{Six-hourly temperatures for 2019.}
\format{A numeric vector.}
\source{
\url{https://www.ecmwf.int/en/era5-land}
}
\usage{
six_hourly_temps
}
\description{
A dataset containing six-hourly 2m air temperature measurements for 2019 from
the ERA-5 Land dataset. Data for a location near Kinlochewe, Scotland
(-5.2, 57.6).
}
\keyword{datasets}
