% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_moisture.R
\name{soil_moisture}
\alias{soil_moisture}
\title{Calculates volumetric soil moisture.}
\usage{
soil_moisture(
  prec,
  evap,
  cn,
  s_depth = 0.5,
  s_min = 0.074,
  s_max = 0.422,
  smi = (s_max + s_min)/2
)
}
\arguments{
\item{prec}{a vector of precipitation values (mm).}

\item{evap}{a vector of evapotranspiration values (mm).}

\item{cn}{run-off curve number (see details).}

\item{s_depth}{depth of the soil layer (m).}

\item{s_min}{residual fraction content (m^3 / m^3).}

\item{s_max}{volumetric water fraction at saturation (m^3 / m^3).}

\item{smi}{initial volumetric soil water fraction.}
}
\value{
volumetric water fraction for each time step.
}
\description{
\code{soil_moisture} calculates volumetric soil water fraction of
soil.
}
\details{
Applies a simple bucket model to calculate the volumetric soil
water fraction. The run-off curve number controls maximum infiltration rates
such that if rainfall exceeds maximum infiltration, surplus water is assumed
not to enter. Maximum infiltration is calculated by applying the United
States Department of Agriculture method for calculating direct runoff.
Runoff curve numbers can be obtained from here:
https://www.wcc.nrcs.usda.gov/ftpref/wntsc/H&H/training/runoff-curve-numbers1.pdf
Adjustments for antecedent moisture condition are automatically applied, so
the supplied value of \code{cn} should be for average soil moisture (AMC II).
\code{s_min} and \code{s_max} values for soil types are given in \code{\link[ecohydrotools:soilparams]{ecohydrotools::soilparams()}}.
}
\examples{
set.seed(100)
rain <- rgamma(365,4) * rbinom(365,1,0.3)
evap <- rep(1.15,365)
cn <- 90
s_min <- 0.074
s_max <- 0.422
s_depth <- 0.5
sm1 <- soil_moisture(rain, evap, 0.9)
sm2 <- soil_moisture(rain, evap, 0.9, 0.1)
plot(sm1, type = "l", col = "blue", ylim = c(s_min, s_max),
     xlab = "Day", ylab = "Soil moisture")
par(new=T)
plot(sm2, type = "l", col = "red", ylim = c(s_min, s_max),
     xlab = "", ylab = "")

}
