% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm}
\alias{ssm}
\title{ssm: Calculate mean summer soil moisture.}
\usage{
ssm(sm, tme, northern_hemisphere = TRUE)
}
\arguments{
\item{sm}{a vector of soil moisture values (e.g. calculated by \code{soil_moisture})}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{sm}
value.}

\item{northern_hemisphere}{a logical indicating whether Spring dates are to
be calculated for the northern (TRUE) or southern (FALSE) hemisphere.}
}
\value{
a single numeric value of the total precipitation during summer.
}
\description{
\code{ssm} calculates average soil moisture content over the summer
period.
}
\details{
add info about the dates used for summer.
}
\examples{
tme <- tmecreate(2010, 1)
evap <- rep(1.15,length(tme))
sm <- soil_moisture(hourly_precip, evap, 0.9)
ssm(sm, tme)
ssm(sm, tme, northern_hemisphere = FALSE)
}
\seealso{
\code{\link[=soil_moisture]{soil_moisture()}} calculates volumetric soil water fraction of
soil.
}
