% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{submonthly}
\alias{submonthly}
\title{submonthly'}
\usage{
submonthly(tme, div = 2)
}
\arguments{
\item{tme}{a POSIXlt object representing calendar dates and times.}

\item{div}{numeric object specifying the value by which to divide total days of each month. Default = 2.}
}
\value{
a vector of numeric values alternating between 0 and 1.
}
\description{
\code{submonthly} is used to create a list representing each measurement per half-month.
}
\details{
If div=2, value changes after the last 'tme' measurement in each half-month.
}
\examples{
tme <- tmecreate(2010, 6)
sm <- submonthly(tme, div = 2)
sm2 <- submonthly(tme, div = 4)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
