% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{summerprecip}
\alias{summerprecip}
\title{summerprecip: Total Summer precipitation}
\usage{
summerprecip(startyear, endyear, precipnc, startday = 152,
  endday = 243, msk, fo)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{precipnc}{full path name of nc file containing precipitation values  and with data extent: -1.25, 358.75, -91.25, 91.25 when converted to raster format.}

\item{startday}{assumed day of year of start of summer in northen hemisphere in non-leap year* (default 1st June).}

\item{endday}{assumed day of year of end of summer in northen hemisphere in non-leap year* (default 31st Aug).}

\item{msk}{optional land/sea mask raster object with sea areas assigned 'NA' and land areas assigned value '1'. Same extent as precipnc.}

\item{fo}{character string indicating the name and file path for the .tif file to be saved.}
}
\value{
A raster object of growing degree days above 10 degrees Celcius during the growing season exported as a .tif file.
}
\description{
\code{summerprecip} calculates the total amount of precipitation falling in summer.
}
\details{
Function has been designed to run with raster data of dimensions 73 x 144.
Seasons are flipped in southern hemisphere. I.e. 1st June (day 152) = day 152+365/2+0.5 = 334 = 1st Dec.
in leap years, 1 day added. Startday and endday should be for northern hemisphere, and are calculated for southern hemisphere within the function.
}
\seealso{
Requires function \code{tsp} to be loaded.

\code{nctoarray}

\code{mtoraster}
}
