% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{summersoilmoist}
\alias{summersoilmoist}
\title{summersoilmoist: Soil moisture content during summer}
\usage{
summersoilmoist(startyear, endyear, soilnc, fi, startday = 152,
  endday = 243)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{soilnc}{full path name of nc file containing soil moisture values for each year and with data extent: -1.25, 358.75, -91.25, 91.25 when converted to raster format.}

\item{fi}{nc file containing soil moisture values for one year.}

\item{startday}{Indicates assumed day of year of start of summer in northern
hemisphere in non-leap year. Default, 1st June.}

\item{endday}{indicates assumed day of year of end of summer in northern
hemisphere in non-leap year. Default, 31st August.}
}
\value{
a matrix of mean summer soil moisture content values over the specified years.
}
\description{
Calculates mean soil moisture content during the 6-month summer period.
}
\details{
Function has been designed to run with raster data of dimensions 73 x 144.
Seasons are flipped in southern hemisphere. I.e. 1st June (day 152) = day 152+365/2+0.5 = 334 = 1st Dec.
in leap years, 1 day added. Startday and endday should be for northern hemisphere, and are calculated for southern hemisphere within the function.
}
\seealso{
\code{\link[=mtoraster]{mtoraster()}}

\code{\link[=nctoarray]{nctoarray()}} to create array of temperature values from an nc file.

Requires function \code{\link[=ssm]{ssm()}} to be loaded.
}
