% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_thresh.R
\name{temp_thresh}
\alias{temp_thresh}
\title{temp_thresh: Calculate the number of temperature values above or below a
threshold.}
\usage{
temp_thresh(temps, tme, thresh, above = TRUE)
}
\arguments{
\item{temps}{a vector of hourly temperatures.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{temps} value.}

\item{thresh}{a single numeric value defining the temperature threshold.}

\item{above}{logical indicating whether values above or below threshold are
to be counted. Default is TRUE.}
}
\value{
a single numeric value indicating the number of temperature values
above or below a threshold.
}
\description{
\code{temp_thresh} is used to calculate the number of input temperature
values that fall above or below a user-defined threshold.
}
\details{
...
}
\examples{
tme <- tmecreate(2019, 1)
plot(hourly_temps~as.POSIXct(tme), type = "l", xlab = "Month",
ylab = "Temperature")
temp_thresh(hourly_temps, tme, thresh = 5)
temp_thresh(hourly_temps, tme, thresh = -1, above = FALSE)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
