% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tmecreate}
\alias{tmecreate}
\title{tmecreate: Creates a \code{POSIXlt} object representing calendar dates and times}
\usage{
tmecreate(years, hourint = 6)
}
\arguments{
\item{years}{a vector of years.}

\item{hourint}{a single numeric value representing the time interval in hours
between dates and times.}
}
\value{
an object of class \code{POSIXlt} with calander dates and times.
}
\description{
\code{tmecreate} is used to create a \code{POSIXlt} object representing
calendar dates and times for the years specified by \code{years} and for time
intervals specified by \code{hourint}.
}
\examples{
head(tmecreate(2010, 24))
tail(tmecreate(2010, 6))
}
