% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp.R
\name{tsp}
\alias{tsp}
\title{tsp: Calculate total summer precipitation.}
\usage{
tsp(prec, tme, northern_hemisphere = TRUE)
}
\arguments{
\item{prec}{a vector of precipitation values.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each
\code{prec} value.}

\item{northern_hemisphere}{a logical indicating whether Spring dates are to
be calculated for the northern (TRUE) or southern (FALSE) hemisphere.}
}
\value{
a single numeric value of the total precipitation during summer.
}
\description{
\code{tsp} calculates total precipitation during the summer season.
}
\details{

}
\examples{
tme <- tmecreate(2010, 1)
tsp(hourly_precip, tme)
tsp(hourly_precip, tme, northern_hemisphere = FALSE)

}
