% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{warna}
\alias{warna}
\title{warna}
\usage{
warna()
}
\value{
If true, returns warning message, "Data spans more than one year. Data aggregated by unique month irrespective of year and one value returned".
}
\description{
Prints a warning message if data span more than one year and indicates method followed.
}
