% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{warnb}
\alias{warnb}
\title{warnb}
\usage{
warnb()
}
\value{
If true, returns warning message, "Data spans more than one year. Calculations performed on all data and single value returned".
}
\description{
Prints a warning message if data span more than one year and indicates method followed.
}
