% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio1}
\alias{bio1}
\title{bio1: Calculates mean annual temperature}
\usage{
bio1(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.
Ignored if method unspecified.}

\item{method}{an optional character string describing the method used to
calculate mean annual temperature. Options are "anuclim", "dailymaxmin" or unspecified
(see details).}
}
\value{
a single numeric value of mean annual temperature.
}
\description{
\code{bio1} is used to calculated mean annual temperature.
}
\details{
If \code{method} is "anuclim" temperatures are aggregated by month, and
spline intepolated to weekly before mean annual temperature is calculated,
replicating the method used by http://www.worldclim.org/. If \code{method} is
"dailymaxmin", daily mean temperatures are calculated as the mean of daily
maximum and minimum temperatures and annual mean calculated from daily
means. Otherwise the mean of \code{temps} is returned. If using \code{anuclim} method
and data span more than one year, data are aggregated by unique month
irrespective of year and one value returned. If using \code{dailymaxmin} method
and data span more than one year, calculations will be performed on all data
and a single value returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio1(temps, tme)
bio1(temps, tme, method = "anuclim")
bio1(temps, tme, method = "dailymaxmin")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
