% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio11}
\alias{bio11}
\title{bio11: Calculates mean temperature of the coldest quarter}
\usage{
bio11(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{An optional character vector describing the method used to calculate
the mean temperature of the coldest quarter. Options  are "anuclim" or unpsecified (see
details).}

\item{prec}{a vector of precipitation values, normally for one year (see
details).}
}
\value{
a single numeric value of mean temperature of the warmest quarter of
the year.
}
\description{
\code{bio11} is used to calculate the mean temperature of the coldest
quarter (three months) of the year
}
\details{
If method is "anuclim", mean monthly temperature values are
calculated and spline interpolated to a weekly time period. Mean temperature of the coldest 13-week period is
determined. If method is left unspecified, mean temperature of the coldest 3-month (91-day) period is
calculated from annual temperature values.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio11(temps, tme)
bio11(temps, tme, method = "anuclim")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
