% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio12}
\alias{bio12}
\title{bio12: Calculates total annual precipitation}
\usage{
bio12(prec, tme, method = "")
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{An optional character string describing the method used to
calculate total annual precipitation. Options are "anuclim" or unspecified (see
details).}
}
\value{
a single numeric value of total annual precipitation.
}
\description{
\code{bio12} is used to calculate total precipitation in the year
}
\details{
If method is "anuclim", monthly precipitation values are spline
interpolated to a weekly time period and the total for each year returned.
Otherwise, all precipitation values for each year are summed.
}
\examples{
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Precipitation")
bio12(prec, tme)
bio12(prec, tme, method="anuclim")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
