% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio13}
\alias{bio13}
\title{bio13: Calculates precipitation of the wettest period}
\usage{
bio13(prec, tme, method = "week")
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{prec} value.}

\item{method}{An optional character string describing how the maximum weekly or
monthly precipitation is calculated. Options are "week" and "month" (see
details).}
}
\value{
a single numeric value of total precipitation in the wettest week or month of the year.
}
\description{
\code{bio13} is used to calculate the precipitation of the wettest
week or month of the year, depending on the time step.
}
\details{
If method is "week", monthly precipitation values are spline interpolated to a
weekly time period and the maximum weekly precipitation is returned. If method
is "month", monthly precipitation values are summed and the maximum monthly
precipitation is returned.

If data span more than one year, data are aggregated by unique month
irrespective of year and one value returned.
}
\examples{
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Precipitation")
bio13(prec, tme, method="week")
bio13(prec, tme, method="month")
bio13(prec, tme)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
