% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio14}
\alias{bio14}
\title{bio14: Calculates precipitation of the driest period}
\usage{
bio14(prec, tme, method = "week")
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{prec} value.}

\item{method}{an optional character string describing how the minimum weekly
or monthly precipitation is calculated. Options include"week", "month" or unspecified
(see details).}
}
\value{
a single numeric value of total precipitation in the driest week or
month of the year.
}
\description{
\code{bio14} is used to calculate the precipitation in the driest
period of the year
}
\details{
If method is "week" or left unspecified, monthly precipitation values
are spline interpolated to a weekly time period and the minimum weekly
precipitation is returned. If method is "month", the minimum monthly
precipitation is returned.
}
\examples{
prec <- (100 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 100)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Precipitation")
bio14(prec, tme)
bio14(prec, tme, method="week")
bio14(prec, tme, method="month")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
