% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio16}
\alias{bio16}
\title{bio16: Calculates precipitation of the wettest quarter}
\usage{
bio16(prec, tme, method = "")
}
\arguments{
\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{prec} value.}

\item{method}{an optional character string describing how precipitation of the wettest
quarter is calculated. Options include "anuclim" or unspecified (see details).}
}
\value{
a single numeric value for precipitation in the wettest quarter of the year.
}
\description{
\code{bio16} is used to calculate the total precipitation of the
wettest quarter of the year
}
\details{
If method is "anuclim", monthly precipitation is spline interpolated
to a weekly time period. Precipitation for each 13-week period is calculated
and total precipitation in the wettest quarter returned. If data span more
than one year, data are aggregated by unique month irrespective of year and
one value returned. Otherwise, precipitation for each three-month (91-day)
period is calculated and total precipitation in the wettest quarter
returned. If data span more than one year, calculations are performed on all
data and a single value returned.
}
\examples{
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Precipitation")
bio16(prec, tme)
bio16(prec, tme, method="anuclim")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
