% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio18.R
\name{bio18}
\alias{bio18}
\title{bio18: Calculate precipitation of the warmest quarter.}
\usage{
bio18(temps, prec, tme1, tme2)
}
\arguments{
\item{temps}{a vector of temperature values, normally for one year (see
details).}

\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}
}
\value{
a single numeric value of total precipitation in the warmest quarter
of the year.
}
\description{
\code{bio18} is used to calculate the total precipitation in the
warmest quarter of the year.
}
\details{
If \code{temps} and \code{prec} vary in their temporal resolution, then they
are harmonised by aggregation (mean for temperature or sum for precipitation).
Then, each data point is treated as the start of a 91 day period for which
summed temperature is calculated. Once the warmest 91 day period is
identified, the total precipitation for the same period is calculated and a
single value returned. If data span more than one year, calculations are
performed on all data and a single value returned.
}
\examples{
# hourly
temps <- 10 * sin(c(0:8759) / (pi * 900)) + rnorm(8760)
tme1 <- tmecreate(2010, 1)
# 6-hourly
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme1 <- tmecreate(2010, 6)
# daily
temps <- 10 * sin(c(0:364) / (pi * 37.5)) + rnorm(365)
tme1 <- tmecreate(2010, 24)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
# hourly
prec <- (10 * sin(c(0:8759) * (pi / -9000)) + rnorm(8760) + 14)
tme2 <- tmecreate(2010, 1)
# 6-hourly
prec <- (10 * sin(c(0:1459) * (pi / -1500)) + rnorm(1460) + 14)
tme2 <- tmecreate(2010, 6)
# daily
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 14)
tme2 <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme2), type = "l", xlab = "Month",
ylab = "Precipitation")
bio18(temps, prec, tme1, tme2)
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
