% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio18}
\alias{bio18}
\title{bio18: Precipitation of the warmest quarter}
\usage{
bio18(temps, prec, tme1, tme2, method = "")
}
\arguments{
\item{temps}{a vector of temperature values, normally for one year (see
details).}

\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec} value.}

\item{method}{an optional character string describing how quarterly
temperature and precipitation are calculated. Options are "anuclim" or unspecified
(see details).}
}
\value{
a single numeric value of precipitation in the warmest quarter of the
year.
}
\description{
\code{bio18} is used to calculate the precipitation in the warmest
quarter of the year
}
\details{
If method is "anuclim", monthly mean temperature and total monthly
precipitation are interpolated to a weekly time period before calculating
mean temperature of each 13-week period. The precipitation in the warmest
quarter is then found. If data span more than one year, data are aggregated
by unique month irrespective of year and one value returned. If method is
unspecified, the mean temperature in each three-month period is calculated
and precipitation in the coldest quarter returned. If data span more than
one year, calculations are performed on all data and single value returned.
}
\examples{
prec <- (100 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 100)
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme1 <- tmecreate(2010, 6)
tme2<- tmecreate(2010, 24)
plot(temps~as.POSIXct(tme1), type = "l", xlab = "Month", ylab = "Temperature")
bio18(temps, prec, tme1, tme2)
bio18(temps, prec, tme1, tme2, method="anuclim")
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
