% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio19}
\alias{bio19}
\title{bio19: Precipitation of the coldest quarter}
\usage{
bio19(temps, prec, tme1, tme2, method = "")
}
\arguments{
\item{temps}{a vector of temperature values, normally for one year (see
details)}

\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec} value.}

\item{method}{an optional character string describing how quarterly mean
temperature and precipitation are calculated. Options are "anuclim" or unspecified
(see details).}
}
\value{
a single numeric value of precipitation in the coldest quarter of the
year.
}
\description{
\code{bio19} is used to calculate the precipitation in the coldest
quarter of the year.
}
\details{
If method is "anuclim", monthly mean temeprature and total monthly
precipitation are interpolated to weekly time period before calculating mean
temperature for each 13-week period. The precipitation in the coldest
quarter is then calculated. If data spans more than one year, data are
aggregated by unique month irrespective of year and one value returned If
method is left unspecified, the mean temperature in each three-month period
is calculated and precipitation in the coldest quarter returned. If data
spans more than one year, calculations are performed on all data and single
value returned.
}
\examples{

prec <- 10 * sin(c(0:364) * (pi / -360)) + (rnorm(365) + 12)
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme1 <- tmecreate(2010, 6)
tme2<- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Precipitation")
bio19(temps, prec, tme1, tme2)
bio19(temps, prec, tme1, tme2, method="anuclim")
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
