% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio2}
\alias{bio2}
\title{bio2: Calculates mean annual diurnal temperature range}
\usage{
bio2(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{an optional character string describing the method used to
mean annual diurnal temperature range. Options are "anuclim" or unspecified (see
details).}
}
\value{
a single numeric value of mean diurnal temperature range.
}
\description{
\code{bio2} is used to calculate the mean annual diurnal range in
temperature (range mean of the maximum-minimum).
}
\details{
If using \code{anuclim} method and data span more than one year, data are
aggregated by unique month irrespective of year and one value returned. If
\code{method} is "anuclim" temperatures are aggregated by month and spline
intepolated to weekly before mean diurnal temperature range is calculated,
replicating the method used by http://www.worldclim.org/. If left
unspecified and time interval is <= daily, the mean difference between the
daily maximum and minimum values is calculated.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio2(temps, tme)
bio2(temps, tme, method = "anuclim")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
