% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio3}
\alias{bio3}
\title{bio3: Calculates isothermality}
\usage{
bio3(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{An optional character string describing the method used to
calculate isothermality. Options include "anuclimmean", "anuclimmaxmin" or
unspecified (see details).}
}
\value{
a single numeric value representing isothermality for each year.
}
\description{
\code{bio3} is used to calculate isothermality (day-to-night
temperature oscillations relative to annual oscillations).
}
\details{
If method is "anuclimmean", bio3 is calculated using "anuclim" method.
Temperatures are aggregated by month and spline intepolated to weekly before
mean diurnal temperature range is calculated, replicating the method used by
http://www.worldclim.org/.

If method is "anuclimaxmin", bio3 is calculated using "anuclimmaxmin" method.
Maximum and minimum monthly temperatues are spline interpolated to a weekly
time period and range is calculated from the maximum and minimum mean weekly
temperature values.

If using method "anuclimmean" or "anuclimmaxmin" and data spans more than one
year, data are aggregated by unique month irrespective of year and one value
returned.

If method is left unspecified, bio3 is calculated using the mean of daily
maximum and minimum temperature. If data spans more than one year, data are
aggregated by unique month irrespective of year and one value returned. If
method is unspecified, bio7 is calculated using the maximum temperature value
for the year.If data spans more than one year, bio7 calculations are performed
on all data and single value returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 24)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio3(temps, tme)
bio3(temps, tme, method = "anuclimmean")
bio3(temps, tme, method = "anuclimmaxmin")

}
\seealso{
\code{\link[=bio2]{bio2()}} and \code{\link[=bio7]{bio7()}} for calculating diurnal and annual temperature ranges.
\code{\link[=tmecreate]{tmecreate()}} for creating a 'POSIXlt' object.
}
