% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio4}
\alias{bio4}
\title{bio4: Calculates temperature seasonality}
\usage{
bio4(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{an optional character string describing the method used to
calculate temperature seasonality. Options are "anuclim" or unspecified (see
details).}
}
\value{
a single numeric value representng annual temperature seasonality.
}
\description{
\code{bio4} calculates the variation in temperature over a given year
as the coeeficient of variation in the mean temperatures
}
\details{
If method is "anuclim" temperatures are aggregated by month and
monthly averages are calculated and spline intepolated to weekly values.
Temperature seasonality is calculated as the standard deviation of weekly
mean temperatures as a percentage of the mean of those temperatures. If
using "anuclim" method and data span more than one year, data are aggregated
by unique month irrespective of year and one value returned. If method is
not specified, calculation is based on the standard deviation of the mean of
all temperatures as a percentage of the mean of all temperatures. If method
is not specified and data span more than one year, calculations will be
performed on all data and a single value returned. For all calculations, the
mean in degrees Kelvin is used.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio4(temps, tme)
bio4(temps, tme, method = "anuclim")
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
