% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio5}
\alias{bio5}
\title{bio5: Calculates maximum temperature of the warmest period of the year}
\usage{
bio5(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{an optional character string describing the method to calculate
maximum temperature in the warmest period. Options are "anuclimmean",
"anuclimmax" or unspecified (see details).}
}
\value{
a single numeric value of maximum temperature (weekly or monthly) for each year.
}
\description{
\code{bio5} is used to calculate the maximum weekly or monthly temperature in each year
}
\details{
If method is "anuclimmean", mean monthly temperatures are spline
interpolated to a weekly time period and the maximum weekly value returned.
If method is "anuclimmax", monthly maximum temperatures are spline
interpolated to a weekly time period and the maximum weekly value returned.
If method is unspecified, the maximum temperature across all values for each year is returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio5(temps, tme)
bio5(temps, tme, method = "anuclimmean")
bio5(temps, tme, method = "anuclimmax")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
