% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio6}
\alias{bio6}
\title{bio6: Calculates minimum temperature of the coldest period of the year}
\usage{
bio6(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{an optional character string describing the method to calculate
minimum temperature. Options are "anuclimmean", "anuclimmin" or unspecified (see
details).}
}
\value{
a single numeric value of minimum temperature for the defiend period.
}
\description{
\code{bio6} is used to calculate the minimum temperature value across
all months of the year.
}
\details{
"anuclimmean" splines mean monthly temperature to weekly time period
across all months and returns the mean minimum weekly temperature. Anuclimmax
splines minimum monthly temperatures to weekly time period and returns the
minimum weekly temperature. If left unspecified, the minimum temperature
across all values of each year is returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio6(temps, tme)
bio6(temps, tme, method = "anuclimmean")
bio6(temps, tme, method = "anuclimmin")
}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
