% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio7}
\alias{bio7}
\title{bio7: Annual temperature range}
\usage{
bio7(temps, tme, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{method}{An optional character string describing the method used to
calculate maximum and minimum temperatures. Options are "anuclimmean",
"anuclimmax" or unspecified (see details).}
}
\value{
a single numeric value of annnual temperature range (maximum-minimum
temperature values).
}
\description{
\code{bio7} is used to calculate the annual range in temperature
}
\details{
If method is "anuclimmean", mean monthly temperatures are spline
interpolated to a weekly time period and range calculated from the maximum
and minimum weeekly temperature values. If method is "anuclimmaxmin",
maximum and minimum monthly temperatues are spline interpolated to a weekly
time period and range is calculated from the maximum and minimum mean weekly
temperature values. If left unspecified, the range is calculated by
subtracting the maximum and minimum temperature alues for each year. To
satisfy the requirements for \code{tme}, a POSIXlt object can be created using
the \code{tmecreate} wrapper function. This calculation should return the value of
bio5(temps, tme)-bio6(temps, tme) when methods remain the same.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
bio7(temps, tme)
bio7(temps, tme, method = "anuclimmean")
bio7(temps, tme, method = "anuclimmaxmin")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
