% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio8}
\alias{bio8}
\title{bio8: Calculates mean temperature of Wettest quarter}
\usage{
bio8(temps, prec, tme1, tme2, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec} value.}

\item{method}{An optional character string describing the methods used to
calculate mean temperature in the wettest quarter. Options are "anuclim" or unspecified
(see details).}
}
\value{
a single numeric value of mean temperature of the wettest quarter of
the year.
}
\description{
\code{bio8} is used to calculate the mean temperature in the wettest
quarter of the year
}
\details{
If method is "anuclim", mean monthly temperature and total monthly
precipitation is calculated and then spline interpolated to a weekly time
period. Precipitation is calculated for each 13-week period and the mean
temperature for the wettest period returned. If data spans more than one
year, data are aggregated by unique month irrespective of year and one value
returned. If method is unspecified, the mean temperature of the wettest
quarter is calculated using all \code{temps} values and precipitation per quarter
is calculated using the time interval for measurements. If data span more
than one year, calculations are performed on all data and a single value
returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme1 <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme1), type = "l", xlab = "Month", ylab = "Temperature")
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 100)
tme2 <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme2), type = "l", xlab = "Month", ylab = "Precipitation")
bio8(temps, prec, tme1, tme2)
bio8(temps, prec, tme1, tme2, method = "anuclim")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
