% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimvars.R
\name{bio9}
\alias{bio9}
\title{bio9: Calculates mean temperature of the driest quarter}
\usage{
bio9(temps, prec, tme1, tme2, method = "")
}
\arguments{
\item{temps}{a vector of temperatures, normally for one year (see details).}

\item{prec}{a vector of precipitation values, normally for one year (see
details).}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps} value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec} value.}

\item{method}{character string describing the method used to calculate mean temperature
of the driest quarter. Options are "anuclim" or unspecified (see details).}
}
\value{
a single numeric value of mean temperature of the wettest quarter of
the year.
}
\description{
\code{bio9} is used to calculate the mean temperature of the driest
quarter of the year
}
\details{
If method is "anuclim", mean monthly temperature values are
calculated and spline interpolated to a weekly time period. Precipitation
values are summed for all months and then spline interpolated to a weekly
time period. Mean temeprature of the driest 13-week period is returned.
Otherwise, annual precipitation values are used to calculate precipitation
in the driest three-month period and mean temperature in this period
returned.
}
\examples{
temps <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
tme1 <- tmecreate(2010, 6)
plot(temps~as.POSIXct(tme), type = "l", xlab = "Month", ylab = "Temperature")
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 12)
tme2 <- tmecreate(2010, 24)
plot(prec~as.POSIXct(tme2), type = "l", xlab = "Month", ylab = "Precipitation")
bio9(temps, prec, tme1, tme2)
bio9(temps, prec, tme1, tme2, method = "anuclim")

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.
}
