% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseason.R
\name{gseason}
\alias{gseason}
\title{gseason: Calculate the period when plant growth is possible.}
\usage{
gseason(
  temps,
  prec,
  evap,
  tme1,
  tme2,
  tme3,
  lon,
  lat,
  lower = 5,
  upper = 35,
  nday = 5,
  daynight = TRUE,
  merid = 0,
  dst = 0
)
}
\arguments{
\item{temps}{a vector of temperature values.}

\item{prec}{a vector of precipitation values.}

\item{evap}{a vector of evapotranspiration values.}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{temps}
value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}

\item{tme3}{a \code{POSIXlt} object representing the date and time of each \code{evap}
value.}

\item{lon}{a numeric value indicating the longitudinal position of data (in
decimal degrees).}

\item{lat}{a numeric value indicating the latitudinal position of data (in
decimal degrees).}

\item{lower}{defines lower temperature limit where plant growth ceases
(degrees Celcius) and growing season is switched off.}

\item{upper}{defines upper temperature limit where plant growth ceases
(degrees Celcius) and growing season is switched off.}

\item{nday}{a single numeric value defining the number of days over which to
smooth the data.}

\item{daynight}{if TRUE, growing season is continuous over 24 hours, if
FALSE, only calculates growing season during the day.}

\item{merid}{an optional numeric value representing the longitude (decimal
degrees) of the local time zone meridian (0 for GMT). Default is 0.}

\item{dst}{an optional numeric value representing the time difference from
the timezone meridian (hours, e.g. +1 for BST if merid = 0). Default is 0.}
}
\value{
an array of binary values indicating growing conditions (1) = yes,
(0) = no.
}
\description{
\code{gseason} calculates total annual growing hours where
temperatures are within defined upper and lower limits and precipitation
exceeds half potential evapotranspiration.
}
\details{
The growing season period is defined as the period where
temperatures are >5 degree Celcius and <35 degree Celcius and precipitation0.5 PET.
}
\examples{
tme <- tmecreate(2010,1)
lon <- -5
lat <- 52
gseason(hourly_temps, hourly_precip, hourly_pev, tme, tme, tme, lon, lat)
gseason(hourly_temps, hourly_precip, hourly_pev, tme, tme, tme, lon, lat,
daynight = FALSE)

}
\seealso{
\code{\link[=gseason_prec]{gseason_prec()}} calculates period where precipitation exceeds half
evapotranspiration.

\code{\link[=gseason_temp]{gseason_temp()}} calculates period where temperature are above lower
limit for plant growth and below upper limit for plant growth.

\code{\link[=gseason_day]{gseason_day()}} calculates day/night time.

\code{\link[=tmecreate]{tmecreate()}} can be used to create a POSIXlt object.
}
