% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseason_prec.R
\name{gseason_prec}
\alias{gseason_prec}
\title{gseason_prec: Calculate the growing season period as determined by
precipitation and evapotranspiration balance.}
\usage{
gseason_prec(prec, evap, tme1, tme2, nday = 28)
}
\arguments{
\item{prec}{a vector of precipitation values.}

\item{evap}{a vector of evapotranspiration values.}

\item{tme1}{a \code{POSIXlt} object representing the date and time of each \code{prec}
value.}

\item{tme2}{a \code{POSIXlt} object representing the date and time of each \code{evap}
value.}

\item{nday}{a single numeric value defining the number of days over which to
smooth the data.}
}
\value{
A time-series object of binary values, where 1 indicates growing
season and 0 indicates not growing season.
}
\description{
\code{gseason_prec} calculates growing season period where
precipitation exceeds half the potential evapotranspiration.
}
\details{
The growing season is defined as the period when precipitation
exceeds half the potential evapotranspiration. If \code{prec} and \code{evap} have
differing temporal resolutions, data are aggregated to the coarser of the
two. More about smoothing....
}
\examples{
tme1 <- tmecreate(2010,24)
tme2 <- tmecreate(2010,24)
prec <- (10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 15)
evap <- (10 * sin(c(0:364) * (pi / +360)) + rnorm(365) + 2)
gseason_prec(prec, evap, tme1, tme2)
gseason_prec(prec, evap, tme1, tme2, nday = 10)

}
\seealso{
\code{\link[=tmecreate]{tmecreate()}} can be used to create a POSIXlt object.
}
