% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gseason_prec}
\alias{gseason_prec}
\title{gseason_prec: Calculates the growing season period as determined by precipitation and evapotranspiration balance}
\usage{
gseason_prec(prec, evap, tme, period = "monthly", surplus = TRUE)
}
\arguments{
\item{prec}{a vector of precipitation values.}

\item{evap}{a vector of evapotranspiration values.}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{evap} value.}

\item{period}{an optional character string defining time period of precipitation and evapotranspiration measurements. Options include "monthly" and "bimonthly". If left unspecified, assumed 'submonthly'.}

\item{surplus}{default = TRUE}
}
\value{
A vector of binary values, where 1 indicates growing season and 0 indicates not growing season.
}
\description{
'gseason_prec' calculates growing season period where precipitation exceeds half the potential evapotranspiration.
}
\details{
The growing season is defined as the period when precipitation exceeds half the potential evapotranspiration.
If period is "monthly", data are aggregated by each month.
If period is "bimonthly", data are aggregated by each half-month.
If surplus = 'TRUE', generates a vector of growing season values for the specified time period.
Requires wrapper functions \code{submonthly} and \code{cumsumseq} to be loaded.
}
\examples{
prec <- 10 * sin(c(0:364) * (pi / -360)) + rnorm(365) + 50
evap <- 10 * sin(c(0:364) * (pi / +360)) + rnorm(365) + 80
tme <- tmecreate(2010, 24)
gsp <- gseason_prec(prec, evap, tme, period = "monthly", surplus = TRUE)
gsp <- gseason_prec(prec, evap, tme, period = "bimonthly", surplus = TRUE)
plot(gsp)

}
\seealso{
the \code{\link[=tmecreate]{tmecreate()}} function can be used to create a POSIXlt object.

\code{submonthly}

\code{cumsumseq}
}
