% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gseason_soilmoist}
\alias{gseason_soilmoist}
\title{gseason_soilmoist: Growing season soil moisture content}
\usage{
gseason_soilmoist(startyear, endyear, ga, sm, msk, fo)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{ga}{array of growing binary values indicating growing (1) or non-growing (0) season.}

\item{sm}{array of soil moisture values for one year.}

\item{msk}{optional land/sea mask raster object with sea areas assigned 'NA' and land areas assigned value '1'.}

\item{fo}{character string indicating the name and file path for the .tif file to be saved.}
}
\value{
A raster object of growing season soil moisture values exported as .tif file.
}
\description{
\code{gseason_soilmoist} calculates the average soil moisture content during the growing season over specified years.
}
\details{
Function has been designed to run with raster data of dimensions 73 x 144.
}
\seealso{
Requires function \code{gssm} to be loaded.

the \code{gseason} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.

mtoraster.

Requires that function 'gssm' is also loaded.
}
