% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gseasonlength}
\alias{gseasonlength}
\title{gseasonlength: Growing season length}
\usage{
gseasonlength(startyear, endyear, gseason)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{gseason}{a 3-dimensional array of growing season values for each year (1 = growing season, 0 = non-growing season).}
}
\value{
A matrix of mean growing season length (number of days) over specified years.
}
\description{
\code{gseasonlength} calculates the average length of the growing season period (in days) across specified years.
}
\details{
Some details required here
}
\seealso{
the \code{\link[=gseason]{gseason()}} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.

\code{\link[=mtoraster]{mtoraster()}} to convert a matrix to a raster object.

\code{\link[=nctarray]{nctarray()}} to create array from an nc file.

Requires function \code{gsl} to be loaded.
}
