% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gseasprec}
\alias{gseasprec}
\title{gseasonprecip: Total precipitation during the growing season}
\usage{
gseasprec(startyear, endyear, precip, gseason)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{precip}{3-dimensional array of precipitation values.}

\item{gseason}{3-dimensional array of values indicating growing season (1) or non-growing season (0).}
}
\value{
A matrix of mean total precipitation values during the growing season over the specified years.
}
\description{
\code{gseasonprecip} calculates the total precipitation during the growing season.
}
\seealso{
Requires function \code{gsp} to be loaded.

the \code{gseason} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.

\code{\link[=mtoraster]{mtoraster()}} to convert output matrix to a raster.

\code{\link[=nctarray]{nctarray()}} to create an array from an nc file.
}
