% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gsmax}
\alias{gsmax}
\title{gsmax: Maximum temperature during the growing season}
\usage{
gsmax(temp, gseason)
}
\arguments{
\item{temp}{a three dimensional array of temperature values.}

\item{gseason}{a three dimensional array of binary values (1 = growing season, 0 = not growing season).}
}
\value{
a matrix of values of maximum growing season temperatures.
}
\description{
Calculates maximum air temperature during the growing season.
}
\examples{
temp <- array(10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460), dim=c(73,144,1460))
tme <- tmecreate(2010, 6)
gs <- gseason_day(tme, 6, 21)
gseason <- array(gs, dim=c(73, 144, 1460))
maxgst <- gsmax(temp, gseason)
}
\seealso{
the \code{gseason} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.
}
