% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gsp}
\alias{gsp}
\title{gsp: Total precipitation during growing season}
\usage{
gsp(prec, gseason)
}
\arguments{
\item{prec}{a three dimensional array of precipitation values.}

\item{gseason}{a three dimensional array of binary values (1 = growing season, 0 = not growing season).}
}
\value{
a matrix of values of total precipitation during the growing season.
}
\description{
\code{gsp} calculates total precipitation during the growing season.
}
\examples{
prec <- array(10 * sin(c(0:1459) * (pi / -1400)) + runif(1460, 0, 10) +10, dim=c(1,1,1460))
tme <- tmecreate(2010, 6)
gs <- gseason_day(tme, 6, 21)
gseason <- array(gs, dim=c(1, 1, 1460))
gsp(prec, gseason)

}
\seealso{
the \code{gseason} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.
}
