% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gssm}
\alias{gssm}
\title{gssm: Soil moisture content during the growing season}
\usage{
gssm(gseason, soilm)
}
\arguments{
\item{gseason}{a three-dimensional array of binary values indicating growing conditions (1) = yes, (0) = no.}

\item{soilm}{a three-dimensional array of fractional soil moisture values.}
}
\value{
a matrix of mean soil moisture values during the growing season.
}
\description{
\code{gssm} calculates mean soil moisture content during the growing season period.
}
\examples{
require(microclima)
tme <- as.POSIXlt(c(0:1459) * 3600 * 6, origin = "2010-01-01 00:00", tz = "GMT")
gs <- gseason_day(tme, 6, 21)
gseason <- array(gs, dim=c(1, 1, 1460))
soilm <- array(runif(1460, 0, 100), dim= c(1,1,1460))
gssm <- gssm(gseason, soilm)

}
\seealso{
the \code{\link[=gseason]{gseason()}} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.
}
