% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{gst}
\alias{gst}
\title{gst: Mean growing season temperature}
\usage{
gst(temp, gseason)
}
\arguments{
\item{temp}{a three-dimensional array of air temperature values.}

\item{gseason}{a three-dimensional array of binary values indicating growing conditions (1) = yes, (0) = no.}
}
\value{
a matrix of mean air temperature during the growing season.
}
\description{
\code{gst} calculates mean temperature during the growing season.
}
\examples{
tme <- as.POSIXlt(c(0:1459) * 3600 * 6, origin = "2010-01-01 00:00", tz = "GMT")
gs <- gseason_day(tme, 6, 21)
gseason <- array(gs, dim=c(1, 1, 1460))
temp <- 10 * sin(c(0:1459) / (pi * 150)) + rnorm(1460)
meangst <- gst(temp, gseason)
}
\seealso{
the \code{gseason} function can be used to create an array of growing conditions (1) = yes, (0) = no accounting for temperature, precipitation and daylight hours.
}
