% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{maxgstemp}
\alias{maxgstemp}
\title{maxgstemp: Maximum temperature during the growing season}
\usage{
maxgstemp(startyear, endyear, temp, gs)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{temp}{a three-dimensional array of temperature values for one year.}

\item{gs}{a three-dimensional array of binary values indicating growing (1) or non-growing (0) season for one year.}
}
\value{
A matrix of mean maximum growing season air temperature values over the specified years.
}
\description{
Calculates the maximum air temperature during the growing season across specified years.
}
\details{
some details needed here
}
\seealso{
\code{\link[=mtoraster]{mtoraster()}} to convert a matrix to a raster object.

\code{\link[=nctarray]{nctarray()}} to create array of temperature values from an nc file.

Requires function \code{\link[=gsmax]{gsmax()}} to be loaded.
}
