% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R, R/ssm.R
\name{ssm}
\alias{ssm}
\title{ssm: Summer soil moisture content}
\usage{
ssm(sm, tme, northern_hemisphere = TRUE)

ssm(sm, tme, northern_hemisphere = TRUE)
}
\arguments{
\item{sm}{a vector of soil moisture values (e.g. calculated by \code{soil_moisture})}

\item{tme}{a \code{POSIXlt} object representing the date and time of each \code{sm}
value.}

\item{northern_hemisphere}{a logical indicating whether Spring dates are to
be calculated for the northern (TRUE) or southern (FALSE) hemisphere.}

\item{soilm}{a three dimensional array of soil moisture values for one year.}

\item{year}{calendar year.}

\item{startday}{assumed day of year of start of summer in northen hemisphere in non-leap year.}

\item{endday}{assumed end of summer. Defaults are 1st June to 31st Aug.}

\item{r}{a raster of same extent as temp coded as 1 for northern hemisphere and 0 for southern hemisphere.}
}
\value{
a matrix of mean summer soil moisture values.

a single numeric value of the total precipitation during summer.
}
\description{
\code{ssm} calculates average soil moisture content over the summer period.

\code{ssm} calculates average soil moisture content over the summer
period.
}
\details{
Seasons are flipped in southern hemisphere. I.e. 1st June (day 152) = day 152+365/2+0.5 = 334 = 1st Dec.
in leap years, 1 day added. Startday and endday should be for northern hemisphere, and are calculated for southern hemisphere within the function.

add info about the dates used for summer.
}
\examples{
soilm <- array(runif(1460, 0, 65), dim= c(73,144,1460))
m <- matrix(1, 73, 144)
r <- raster(m, crs="+init=epsg:4326")
extent(r) <- c(-1.25, 358.75, -91.25, 91.25)
enorth<-extent(-1.25,358.75,0,91.25)
esouth<-extent(-1.25,358.75,-91.25,0)
rn<-crop(r,enorth) * 0 + 1
rs<-crop(r,esouth) * 0
r<-mosaic(rn,rs,fun=mean)
ssm(soilm, 2010, startday = 152, endday = 243, r)

tme <- tmecreate(2010, 1)
evap <- rep(1.15,length(tme))
sm <- soil_moisture(hourly_precip, evap, 0.9)
ssm(sm, tme)
ssm(sm, tme, northern_hemisphere = FALSE)
}
\seealso{
\code{\link[=soil_moisture]{soil_moisture()}} calculates volumetric soil water fraction of
soil.
}
