% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{summerprecip}
\alias{summerprecip}
\title{summerprecip: Total Summer precipitation}
\usage{
summerprecip(startyear, endyear, precipnc, startday = 152,
  endday = 243)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{precipnc}{full path name of nc file containing precipitation values  and with data extent: -1.25, 358.75, -91.25, 91.25 when converted to raster format.}

\item{startday}{assumed day of year of start of summer in northen hemisphere in non-leap year* (default 1st June).}

\item{endday}{assumed day of year of end of summer in northen hemisphere in non-leap year* (default 31st Aug).}
}
\value{
A matrix of mean growing degree days above 10 degrees Celcius during the growing season over the specified years.
}
\description{
\code{summerprecip} calculates the total amount of precipitation falling in summer.
}
\details{
Function has been designed to run with raster data of dimensions 73 x 144.
Seasons are flipped in southern hemisphere. I.e. 1st June (day 152) = day 152+365/2+0.5 = 334 = 1st Dec.
in leap years, 1 day added. Startday and endday should be for northern hemisphere, and are calculated for southern hemisphere within the function.
}
\seealso{
Requires function \code{\link[=tsp]{tsp()}} to be loaded.

\code{\link[=mtoraster]{mtoraster()}} to convert output matrix to a raster.

\code{\link[=nctarray]{nctarray()}} to create an array from an nc file.
}
