% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physclimvars.R
\name{summersoilmoist}
\alias{summersoilmoist}
\title{summersoilmoist: Soil moisture content during summer}
\usage{
summersoilmoist(startyear, endyear, soilnc, fi, startday = 152,
  endday = 243, msk, fo)
}
\arguments{
\item{startyear}{earliest calender year to be considered in calculations.}

\item{endyear}{latest calender year to be considered in calculations.}

\item{soilnc}{full path name of nc file containing soil moisture values for each year and with data extent: -1.25, 358.75, -91.25, 91.25 when converted to raster format.}

\item{fi}{nc file containing soil moisture values for one year.}

\item{startday}{Indicates assumed day of year of start of summer in northern
hemisphere in non-leap year. Default, 1st June.}

\item{endday}{indicates assumed day of year of end of summer in northern
hemisphere in non-leap year. Default, 31st August.}

\item{msk}{raster object of land-sea mask with sea areas assigned 'NA' and land areas assigned value '1'.}

\item{fo}{character string indicating the name and file path for the .tif file to be saved.}
}
\value{
a raster object of mean summer soil moisture content exported as a .tif file.
}
\description{
Calculates mean soil moisture content during the 6-month summer period.
}
\details{
Function has been designed to run with raster data of dimensions 73 x 144.
Seasons are flipped in southern hemisphere. I.e. 1st June (day 152) = day 152+365/2+0.5 = 334 = 1st Dec.
in leap years, 1 day added. Startday and endday should be for northern hemisphere, and are calculated for southern hemisphere within the function.
}
\seealso{
\code{mtoraster}.

\code{nctoarray} to create array of temperature values from an nc file.

Requires function \code{ssm} to be loaded.
}
